Split gpsdio.cli.commands into individual documents, changed variable name in gpsdio insp, and small improvement for GZIP driver.
