+++ gpsdio/drivers.py
@@ -9,6 +9,7 @@ import bz2
 import gzip
 import logging
 from pkg_resources import iter_entry_points
+import sys
 
 import msgpack
 import newlinejson as nlj
@@ -185,7 +186,10 @@ class GZIP(_BaseCompressionDriver):
 
     def open(self, path, mode='r', **kwargs):
 
-        if isinstance(path, six.string_types):
+        # TODO: There's probably a workaround for this?  Not critical but would be cool.
+        if path == sys.stdin:
+            raise TypeError("GZIP can't read directly from stdin")
+        elif isinstance(path, six.string_types):
             return gzip.open(path, mode=mode, **kwargs)
         elif isinstance(path, gzip.GzipFile):
             return path
